Join all Enable and Overcurrent signals
I think this is the problem that has been bothering me for several weeks. When I replaced the TPS2044B with the TPS2041B, I thought I could use any Enable and any Overcurrent signal, but in that configuration the hub would refuse to power on the downstream ports. After testing this with a breadboard, I think this is the solution.